Add another pass for variable resolution
This adds a new Visitor implementation which ensures we have lexical
scoping for variables.

We want the language definition to be tightened up a little.

A variable usage refers to the preceding declaration with the same name
in the innermost scope that encloses the expression where the variable
is used.

This change fixes this program:

```
var a = "global";
{
  fun showA() {
    print a;
  }

  showA();
  var a = "block";
  showA();
}
```

Before:

```
global
block
```

After:

```
global
global
```