Add index-based variable access
Our resolver calculates which environment the variable is found in, but
it's still looked up by name in that map. A more efficient environment
representation would store local variables in an array and look them up
by index.

Extend the resolver to associate a unique index for each local variable
declared in a scope. When resolving a variable access, look up both the
scope the variable is in and its index and store that. In the
interpreter, use that to quickly access a variable by its index instead
of using a map.

I should probably write some benchmarks to see what the benefit of this
approach is. It took a bit of thinking about what needed to change. The
globals are still in a map. But this approach goes from (hopefully!)
constant-time hashmap to an offset lookup of a pointer. That's
intuitively quicker than hashing, then access the right bucket, then
test each item in the bucket.