Add deleting from hash tables
We use a tombstone to mark when an bucket has been removed. Our
tombstone is a `NULL` key and a `true` value.

But we also need to change how we find entries. A tombstone is half an
entry. A later bucket could contain an entry that hashes to an earlier
bucket, and at insertion time, linear probing meant that it ended up 2
buckets along.

So say we hashed to bucket `n`. `n` and `n+1` was already full. So we
store it in `n+2` by linear probing to find the first subsequent
available bucket. Later on, we delete and tombsone the contents of
`n+1`.

When we go to find something that hashes to `n`, we check `n`. That
contains an entry, but not something with the right key. We check `n+1`.
That contains a tombstone. We need to carry on linear probing to
discover that `n+2` contains the right key.