Treat count as entries + tombstones
If we treat tombstones as full buckets, then we might artificially
inflate the load factor and end up with a larger array than needed.

On the other hand, if we treat tombstones like empty buckets, then we
risk not having any empty buckets to terminate a lookup. This is worse
than a few extra array slots! So for load factor, we consider tombstones
to be full buckets.

This means that we increment the count during insertion only if the new
entry goes in an entirely empty (not half-empty!) bucket.

When we resize, we don't copy tombstones over, so we need to recalculate
the count during a resize operation.