Use string interning for hash table keys
We want to ensure that "key" is the same as "key" when testing. This
fixes lookup so that it's not using `==` for comparing strings. That
would work if the string pointed to the same string in memory, but
that's probably not the desired behaviour.

Instead, intern strings and ensure we have no duplicate strings in
memory with the same characers.