Add support for assigning global variables
Some complexity here.

```
a * b = c + d;
```

That should be a syntax error. For it to be a syntax error, the parser's
precedence handling needs to take account of the surrounding context.

In particular, precedence should be low enough to allow assignment.

That knowledge is only used in the `variable` function. But because of
our table-driven approach, it ripples to ParseFn and all
implementations.