Add support for local variables
Tnis change adds compiling and executing local variable declarations.

This means adding support for block statements. Entering a block begins
a scope for local variables, and leaving a block ends a scope.

Here, we're trying to have a much more performant implementation for
local variables than creating a hash map for each scope. Instead, we
are going to have local variables on the stack with indexed access.

We add a new Compiler struct which tracks locals and scope depth. We are
being lazy and declaring this as global rather than passing it along the
call stack. Passing one around would allow us to make a multi-threaded
implementation.

The compiler thus has a hard limit on the number of locals allowed.
Because we're using a single byte to index the local slots, the VM only
supports up to 256 locals in scope at one time.

We support shadowing, but not redeclaration of locals.