Split variable declaration into 2 phases
Permits this sort of oddness:

```
{
  var a = "outer";
  {
    var a = a;
  }
}
```

So we have a variable in the "declare uninitialised" and "ready for use"
states.

We've been frugal with memory here by overloading the usage of depth.
We could instead have added another field to local, but C programming,
man!

Most of the work to support locals is in the compiler. This means that
runtime is cheaper and faster. That's why we have a VM!