Add compilation of `if` statements
This effectively introduces GOTO into our VM. We move the stack pointer
`ip` around to jump to different locations in the program.

This involves a couple of steps:

1. emit a jump opcode
2. compile the `then` statements
3. patch the jump opcode location now that we know how much room the
   `then` statements require

We allow a short (16-bit unsigned integer) for the jump, so we support
skipping up to 65,535 bytes of code. That should be plenty.