Add support for `else` statements
After we compile the `then` branch, we look for an `else` branch.

As part of this change, we clean up the condition value that was left on
the stack as part of the changes for `if` statements. We need to pop
that off regardless of whether there is an `else` branch. The ordering
is:

```
      condition expression
/---- OP_JUMP_IF_FALSE
|     OP_POP
|     then branch
| /-  OP_JUMP
\-+-> OP_POP
  |   else branch
  \-> subsequent code
```

You can see it always executes the condition, optionally jumps, then
pops the condition before executing the `then` or `else` branch.