Add support for logical `and`
This reuses the JUMP machinery introduced for `if` statements.

```
    left operand expression
/-- OP_JUMP_IF_FALSE
|   OP_POP
|   right operand expression
\-> subsequent code...
```

If the left-hand expression is falsey, then we know the entire `and`
expression will be falsey and we leave that value on the stack as the
result of the expression.

Otherwise, we discard the left-hand value and evaluate the right-hand
expression, which becomes the result of the whole `and` expression.