Use functions in the compiler and VM
This change is a refactoring which adds a lot of scaffolding to support
adding compiling and calling functions.

We change the compiler to compile functions rather than chunks.

We introduce call frames to the VM, which track allocating local
variables, and the return address from functions. The VM now needs to
work with the current call frame's instruction pointer rather than the
VM `ip`.