Handle compiling multiple nested functions
Add a linked list of Compiler references. When we start compiling a
function declaration, we create a new Compiler, initialize it and set it
to be the current Compiler. As we compile the function body, bytecode is
emitted into the chunk owned by the new Compiler's function. At the end
of the function block, we call `endCompiler` which returns the new
compiled function object to be stored in the surrounding function's
constant table.