Change to use closures everywhere
An ObjClosure wraps each ObjFunction. We do this even if the function
doesn't actually close over and capture any surrounding local variables.
This is slightly wasteful (and inefficient!) but it simplifies the VM
because we have a consistent approach.

The compiler needs to change to emit a new OP_CLOSURE opcode instead of
OP_CONSTANT to load a function from the surrounding function's constant
table.

The VM needs to change to handle that new opcode. We never call a bare
ObjFunction any more. Instead, we call an ObjClosure.

Our fib script has got 10% slower though. Presumably from extra pointer
chasing - I haven't profiled it.

```
% ./bin/clox input.lox
9.22746e+06
1.11225
```