Add a subset of working closures
As long as all of the variables remain on the stack, we have working
closures with this change. Closing that gap will be done in the next
commit!

We are using the same solution described for Lua. We are keeping the
constraint of having a single-pass compiler, so we want something that
works with that.

This change introduces a level of indirection called an 'upvalue'. An
upvalue refers to a local variable in an enclosing function. Every
closure maintains a dynamic array of upvalues; one for each surrounding
local variable that the closure uses.

The upvalue points back into the stack to where the variable it captured
lives. When the closure needs to access a closed-over variable, it goes
through the corresponding upvalue to reach it. When a function
declaration is first executed and we create a closure for it, the VM
creates the array of upvalues and wires them up to "capture" the
surrounding local variables that the closure needs.

Upvalues serve as the layer of indirection needed to continue to find a
captured local variable even after it moves off the stack.

We want to do as much work as possible during compilation, so that
execution is simple and fast.

We had the object model and opcode changes needed. The complexity is
added to the compiler, which has to recursively resolve upvalues to
ensure values that are multiple scopes outside of a function are
correctly resolved.

The OP_CLOSURE instruction is unique in that it has a variably sized
encoding. For each upvalue the closure captures, there are two
single-byte operands. Each pair of operands specifies what that upvalue
captures. If the first byte is one, it captures a local variable in the
enclosing function. If zero, it captures one of the function’s upvalues.
The next byte is the local slot or upvalue index to capture.