Add closed upvalues
A key feature of closures is that they hold on to the variable as long
as needed, even after the function that declares the variable has
returned.

```
fun outer() {
  var x = "outside";
  fun inner() {
    print x;
  }

  return inner;
}

var closure = outer();
closure();
```

Variables in closures don't have stack semantics. So we need to hoist
them off the stack and onto the heap when the function where they were
declared returns. Closed-over variables need to move to the heap if the
closure outlives the function where the captured variable is declared.

We'll use 'open upvalue' to refer to an upvalue that points to a local
variable still on the stack. When a variable moves to the heap, we are
*closing* the upvalue and the result is, naturally, a "closed upvalue".
The two questions we need to answer are:

1. Where on the heap does the closed-over variable go?
2. When do we close the upvalue?

The answer to the first question is easy. We already have a convenient
object on the heap that represents a reference to a variable—ObjUpvalue
itself. The closed-over variable will move into a new field right inside
the ObjUpvalue struct. That way we don’t need to do any additional heap
allocation to close an upvalue.

The second question is straightforward too. As long as the variable is
on the stack, there may be code that refers to it there, and that code
must work correctly. So the logical time to hoist the variable to the
heap is as late as possible. If we move the local variable right when it
goes out of scope, we are certain that no code after that point will try
to access it from the stack. After the variable is out of scope, the
compiler will have reported an error if any code tried to use it.