Initial markRoots implementation
This is very much the low-hanging fruit, obvious roots. What is a root?

A root is any object that the VM can reach directly without going
through a reference in some other object.

Reachability is defined as:

* All roots are reachable
* Any object referred to from a reachable object is itself reachable

The algorithm that we're going to implement is:

1. Starting with the roots, traverse through object references to find
   the full set of reachable objects.
2. Free all objects not in that set.

Most roots are local variables or temporaries sitting right in the VM's
stack, so we start by walking that.

The other main source of roots are the global variables. Those live in a
hash table owned by the VM.