Mark less obvious roots
The VM maintains a separate stack of CallFrames. Each CallFrame contains
a pointer to the closure being called. The VM uses those pointers to
access constants and upvalues, so those closures need to be kept around
too.

The open upvalue list is another set of values that the VM can directly
reach.