Add markCompilerRoots
A garbage collection can begin during *any* allocation - not just when
the user's program is running. The compiler itself periodically grabs
memory from the heap for literals and the constant table. If the GC runs
while we're in the middle of compiling, then any values the compiler
directly accesses need to be treated as roots too.

We put this into a separate function to keep it cleanly separated from
the rest of the VM.