Trace references
Once we have marked all the roots, we can do this next phase. This is to
trace the references.

This is straightforward translation of the textual algorithm. We pop
gray objects off the stack until it's empty. For each gray object, we
traverse the references and then mark it black. Traversing an object's
references may turn up new white objects that get marked gray and added
to the stack. So this function swings back and forth between turning
white objects gray and gray objects black, gradually advancing the
entire wavefront forward.

We also need to detect cycles. References between objects are directed,
but that doesn't mean they're acyclic! It's entirely possible to have
cycles of objects. When that happens, we need to ensure our collector
doesn't get stuck in an infinite loop as it continually re-adds the same
series of objects to the gray stack.

If the object is already marked, we don't mark it again and thus don't
add it to the gray stack. This ensures that an already-gray object is
not redundantly added and that a black object is not inadvertently
turned back to gray. In other words, it keeps the wavefront moving
forward through only the white objects.