Add sweep phase to GC
At this point, we have a linked list of objects in the heap. We walk
that list, checking the mark bits. If an object is marked (black), we
unmark it so it can be checked in the next GC cycle, leave it alone and
continue past it. If it's unmarked (white), we unlink it from the list
and free it.