Add GC for the string pool
We can't allow the string table to grow unbounded because we only have
finite memory on a computer. We also need to make sure the string table
doesn't have dangling pointers to freed memory.

What we need is a way of detecting whether strings are reachable, and
freeing both the referenced object and the dangling reference.

This is knowmn as a weak reference - https://en.wikipedia.org/wiki/Weak_reference

To remove references to unreachable strings, we need to know which
strings are unreachable. We don't know that until after the mark phase
has completed. But we can't wait until after the sweep phase is done
because by then the objects - and their mark bits - are no longer around
to check. So the right time is exactly between the marking and sweeping
phases.

The string intern table uses only the key of each entry — it's basically
a hash set not a hash map. If the key string object's mark bit is not
set, then it is a white object that is moments from being swept away. We
delete it from the hash table first and thus ensure we won't see any
dangling pointers.