Adjust GC frequency
We have a stop-the-world garbage collector. We don't want to stop the
world too often, but we also don't want the world to be stopped for too
long. When the world is stopped, our throughput is zero and no useful
work happens from the user's perspective. If the world is stopped for a
long time, then our latency is high.

If we wait a long time before we run the collector, then a large number
of dead objects will accumulate. That leads to a very long pause while
the collector runs, and thus high latency. So, clearly, we want to run
the collector really frequently.

But every time the collector runs, it spends some time visiting live
objects. That doesn't really do anything useful (aside from ensuring
that they don't incorrectly get deleted). Time visiting live objects is
time not freeing memory and also time not running user code. If you run
the GC really frequently, then the user's program doesn't have enough
time to even generate new garbage for the VM to collect. The VM will
spend all of its time obsessively revisiting the same set of live
objects over and over, and throughput will suffer. So, clearly, we want
to run the collector really infrequently.

_exit Princess Bride_

We are trying to get something in the middle here. We track the heap
size, and the threshold for when we next do a GC. Every time we allocate
or free memory, we adjust the counter. When the total crosses the
threshold, we run the collector.

The initial choices for threshold here are arbitrary. We'd need more
data from real-world usage to define limits.