Fix GC bug with constant table
The constant table each chunk owns is a dynamic array. When the compiler
adds a new constant to the current function's table, that array may need
to grow. The constant itself may also be some heap-allocated object like
a string or a nested function.

The new object being added to the constant table is passed to
`addConstant()`. At that moment, the object can be found only in the
parameter to that function on the C stack. That function appends the
object to the constant table. If the table doesn't have enough capacity
and needs to grow, it calls `reallocate()`. That in turn triggers a GC,
which fails to mark the new constant object and thus sweeps it right
before we have a chance to add it to the table. Crash.

The fix is to push the constant onto the stack temporarily. Once the
constant table contains the object, we pop it off the stack.