Fix GC bug with interned strings
All strings are interned in clox, so whenever we create a new string, we
also add it to the intern table. Since the string is brand new, it isn't
reachable anywhere. And resizing the string pool can trigger a
collection.

Again, we go ahead and stash the string on the stack first. Once it's
safely in the table, we pop it back off.