Fix GC bug in string concatenation
For numbers, `OP_ADD` pops the two operands off the stack, computes the
result, and pushes that new value back onto the stack. That's perfectly
safe.

But concatenating two strings requires allocating a new character array
on the heap, which can in turn trigger a GC. Since we've already popped
the operand strings by that point, they can potentially be missed by the
mark phase and get swept away. Instead of popping them off the stack
eagerly, we peek them.

That way, they are still hanging out on the stack when we create the
result string. Once that's done, we can safely pop them off and replace
them with the result.