Add faster method invocations
The language semantics define a method invocation as two operations -
accessing the method and then calling the result. Implementing it like
that in the VM has a significant cost. Every single time a Lox program
accesses and invokes a method, the runtime heap allocates a new
`ObjBoundMethod`, initializes its fields, then pulls them right back
out. Later, the GC has to spend time freeing all of those ephemeral
bound methods.

The common case in programs is to access a method and immediately call
it. Since we can recognise this sequence at compile time, we can emit a
new specialized instruction that optimizes the method call.

This is a classic VM optimization technique. Define a new single
instruction called a **superinstruction** that fuses those into a single
instruction with the same behavior as the entire sequence.

There are edge cases that we need to guard against though.

```
class Oops {
  init() {
    fun f() {
      print "not a method";
    }

    this.field = f;
  }
}

var oops = Oops();
oops.field();
```

The last line looks like a method call, so the compiler emits the new
OP_INVOKE opcode. But it's a field access which returns a function which
is then invoked. So we have a guard condition in `invoke` which checks
if this is field access and dispatches appropriately.