Faster hash table probing
Profiling shows that (on some benchmarks), the VM spends 72% of time in
`tableGet()`. Looking at that function, it's mostly a wrapper around
`findEntry()`. Drilling down into that, most of the time is spent on

```
uint32_t index = key->hash % capacity;
```

It turns out the modulo operator is really slow.

In the general case, it's really hard to re-implement a fundamental
arithmetic operator in user code in a way that's faster than what the
CPU itself can do. After all, our C code ultimately compiles down to the
CPU's own arithmetic operations. If there were tricks we could use to go
faster, the chip would already be using them.

However, we can take advantage of the fact that we know more about our
problem than the CPU does. We use modulo here to take a key string's
hash code and wrap it to fit within the bounds of the table's entry
array. That array starts out at eight elements and grows by a factor of
two each time. We know - and the CPU and C compiler do not - that our
table's size is always a power of two.

A faster way to calculate the remainder of a number modulo a power of
two is bit masking.

229 11100101              229 11100101
MOD                       AND
 64 01000000               63 00111111
---                       ---
 37 00100101               37 00100101

You can calculate a number modulo any power of two by simply AND-ing it
with that power of two minus one.

We apply the same optimisation to `tableFindString()` as well.

```
> cat bench.lox
class Zoo {
  init() {
    this.aardvark = 1;
    this.baboon   = 1;
    this.cat      = 1;
    this.donkey   = 1;
    this.elephant = 1;
    this.fox      = 1;
  }
  ant()    { return this.aardvark; }
  banana() { return this.baboon; }
  tuna()   { return this.cat; }
  hay()    { return this.donkey; }
  grass()  { return this.elephant; }
  mouse()  { return this.fox; }
}

var batches = 0;
var zoo = Zoo();
var sum = 0;
var start = clock();
while (clock() - start < 5) {
  for (var i = 0; i < 10000; i = i + 1) {
    sum = sum + zoo.ant()
            + zoo.banana()
            + zoo.tuna()
            + zoo.hay()
            + zoo.grass()
            + zoo.mouse();
  }
   batches = batches + 1;
}

print sum;
print batches;
```

Before:

```
> ./bin/clox bench.lox
6.684e+07
1114
```

After:

```
> ./bin/clox bench.lox
1.2768e+08
2128
```

The book talks making the VM twice as fast for something like the above
benchmark. It looks like I have a similar result here.