Use NaN Boxing to improve performance
On a 64-bit machine, our `Value` type takes up 16 bytes. The struct has
two fields, a type tag and a union for the payload. The largest fields
in the union are an Obj pointer and a double, which are both 8 bytes. To
keep the union field aligned to an 8-byte boundary, the compiler adds
padding after the tag too:

 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6
 \_____/ \_____/ \_____________/
   |        |          |
4-byte tag  |      8-byte union
            |
       4-byte padding

That's pretty big. If we can cut that down, then the VM could pack more
values into the same amount of memory. Most computers have plenty of RAM
these days, so the direct memory savings aren't a huge deal. But a
smaller representation means more `Value`s fit in a cache line. That
means _fewer_ cache misses, which affects speed.

Starting with a 64-bit double as defined in
https://en.wikipedia.org/wiki/IEEE_754.

The spec carves out a special case exponent. When all of the exponent
bits are set, then instead of just representing a really big number, the
value has a different meaning. These values are “Not a Number” (hence,
**NaN**) values. They represent concepts like infinity or the result of
division by zero.

_Any_ double whose exponent bits are all set is a NaN, regardless of the
mantissa bits. That means there's lots and lots of different NaN bit
patterns. IEEE 754 divides those into two categories. Values where the
highest mantissa bit is 0 are called **signalling NaNs**, and the others
are **quiet NaNs**. Signalling NaNs are intended to be the result of
erroneous computations, like division by zero. A chip may detect when
one of these values is produced and abort a program completely. They may
self-destruct if you try to read one.

Quiet NaNs are supposed to be safer to use. They don't represent useful
numeric values, but they should at least not set your hand on fire if
you touch them.

Every double with all of its exponent bits set and its highest mantissa
bit set is a quiet NaN. That leaves 52 bits unaccounted for. We'll avoid
one of those so that we don't step on Intel's "QNaN Floating-Point
Indefinite" value, leaving us 51 bits. Those remaining bits can be
anything. We're talking 2,251,799,813,685,248 unique quiet NaN bit
patterns.

This means a 64-bit double has enough room to store all of the various
different numeric floating-point values and also has room for another
51 bits of data that we can use however we want. That's plenty of room
to set aside a couple of bit patterns to represent Lox's nil, true, and
false values. But what about Obj pointers? Don't pointers need a full 64
bits too?

Fortunately, we have another trick up our other sleeve. Yes, technically
pointers on a 64-bit architecture are 64 bits. But, no architecture I
know of actually uses that entire address space. Instead, most widely
used chips today only ever use the low 48 bits. The remaining 16 bits
are either unspecified or always zero.

If we've got 51 bits, we can stuff a 48-bit pointer in there with three
bits to spare. Those three bits are just enough to store tiny type tags
to distinguish between nil, Booleans, and Obj pointers.

That's NaN boxing. Within a single 64-bit double, you can store all of
the different floating-point numeric values, a pointer, or any of a
couple of other special sentinel values. Half the memory usage of our
current Value struct, while retaining all of the fidelity.

What's particularly nice about this representation is that there is no
need to convert a numeric double value into a “boxed” form. Lox numbers
are just normal, 64-bit doubles. We still need to check their type
before we use them, since Lox is dynamically typed, but we don't need to
do any bit shifting or pointer indirection to go from "value" to
"number".

For the other value types, there is a conversion step, of course. But,
fortunately, our VM hides all of the mechanism to go from values to raw
types behind a handful of macros. Rewrite those to implement NaN boxing,
and the rest of the VM should just work.

Here, we use a single unsigned 64-bit integer to store a `Value` and
conditional compilation in C to ensure we keep the old representation if
we need to support architectures where this optimization won't work.

Converting a `double` to a `Value` and back might look weird. The way to
treat a series of bytes as having a different type without changing
their value at all is `memcpy()`? This looks horrendously slow: Create a
local variable. Pass its address to the operating system through a
ayscall to copy a few bytes. Then return the result, which is the exact
same bytes as the input. Thankfully, because this is the supported idiom
for type punning, most compilers recognize the pattern and optimize away
the `memcpy()` entirely.

On machines where this Nan Boxing (or NaN Tagging) works, this seems to
give a 10% improvement in performance.