Add Sticky Immix sketch for rlox
See http://www.cs.utexas.edu/users/speedway/DaCapo/papers/immix-pldi-2008.pdf

We are adding a memory management scheme based on Immix.

The first thing we want to add is a way of writing objects into our
block of raw memory. We are going to do bump allocation, where we
write objects in sequential order in memory, and bump the pointer to the
space after the last object.

After it's been used for a while, we will have allocated and freed
objects. The block of memory will be fragmented, with gaps in between
live objects. `BlockMeta` is used to keep track of these holes, so that
we can write into them.